Move `RichSbtClient` to its own file
